+++ test/integration/active_job/sidekiq_adapter/preconfigured_perform_now_test.rb
@@ -1,0 +1,47 @@
+require 'test_helper'
+
+module TestActiveJob
+  module SidekiqAdapter
+    class PreconfiguredPeroformNowTest < ActiveSupport::TestCase
+
+      test 'update succes update user when tenant account match' do
+        under_tenant(amazon) do
+          assert_no_changes 'MultiTenantSupport.current_tenant' do
+            UserNameUpdateJob.set(queue: :integration_tests).perform_now(bezos)
+          end
+
+          assert_equal "Jeff Bezos UPDATE", bezos.reload.name
+        end
+      end
+
+      test 'fail to update user when tenant account is missing' do
+        under_tenant nil do
+          assert_no_changes 'MultiTenantSupport.current_tenant' do
+            assert_raise(MultiTenantSupport::MissingTenantError) do
+              UserNameUpdateJob.set(queue: :integration_tests).perform_now(bezos)
+            end
+          end
+        end
+
+        under_tenant(amazon) do
+          assert_equal "Jeff Bezos", @bezos.reload.name # Does not change
+        end
+      end
+
+      test 'fail to update user when tenant account is not match' do
+        under_tenant(apple) do
+          assert_no_changes 'MultiTenantSupport.current_tenant' do
+            assert_raise(MultiTenantSupport::InvalidTenantAccess) do
+              UserNameUpdateJob.set(queue: :integration_tests).perform_now(bezos)
+            end
+          end
+        end
+
+        under_tenant(amazon) do
+          assert_equal "Jeff Bezos", @bezos.reload.name # Does not change
+        end
+      end
+
+    end
+  end
+end